VLAN Support
The real staging environment will use all the VLANs we currently have in production.
This updates the network portion of provisioning to use OpenVSwitch rather than just
a plain linux bridge.  This allows libvirt networks to be configured which act as
tagged access ports.  In a development situation we can add a vlan device to the
vswtich to act as the preseed server, in production we just add eth1 which will
automatically trunk packets to the upstream physical switch.