Networking Improvements
Updates the host constructor to consume an ordered list of subnet-ipv4 tuples.  The main
gain here is that we can provision the public interface of the gateway node so that
guests calling in from cloud.example.com across the public internet can access the
services.  As such I've hacked in a network interfaces fragment for the second NIC,
added in a NAT rule in the network initialization script to handle internet traffic
from this subnet (which is now the default route).