Crazy Commit
Where to start?  Hypervisor host and network control is now farmed off to distributed
staging-hypervisor-agent processes so all that jazz goes via a REST API.  Generic REST
functionality is now shared with a client to interact with a new preseed server which
provides ephemeral preseed management and rendering functionality.  The main codes is
split up into a separate module to make things more maintainable and less inclined to
use globals.