Separate Managers from Classes
Most classes depend on managers not the actually classes themselves, which can lead to
circular dependencies when say a host needs to use a nameserver to create an A record,
but also needs to get the host address of the nameserver.  This separates these concerns
so the classes aren't so tightly coupled.